pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel", "pandas"]
build-backend = "setuptools.build_meta"

[project]
name = "bridging"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["mdtraj", "numpy", "openmm", "pandas", "pdbfixer-wheel>=1.12.4", "python-docx", "requests"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
